"4:45pm.
```
Starting timing for: iteration 0
The time was 00:00:02.9140790 for: iteration 0
Training: 0.1836399253414
Testing: 0.113177325552664(9620/9984)
Starting timing for: iteration 1
The time was 00:00:02.9030891 for: iteration 1
Training: 0.0643514293120601
Testing: 0.095317505673171(9671/9984)
Starting timing for: iteration 2
The time was 00:00:02.8992291 for: iteration 2
Training: 0.0305183720871961
Testing: 0.0884926123858(9712/9984)
Starting timing for: iteration 3
The time was 00:00:02.9009701 for: iteration 3
Training: 0.0138830112601316
Testing: 0.0851460069254524(9749/9984)
Starting timing for: iteration 4
The time was 00:00:02.9041153 for: iteration 4
Training: 0.00584066336190688
Testing: 0.0820976331742713(9754/9984)
```

2.9 huh? Let me put those symmetric multiplies in.

```
                inl reproject a b ret =
                    inb x = s.CudaBlas.symm .Left .Lower one a.basic b.basic |> CudaAux.temporary
                    ret x
```

Let me just change this for starters.

4:55pm.

```
Starting timing for: iteration 0
The time was 00:00:03.0417623 for: iteration 0
Training: 0.183482415679619
Testing: 0.115261024058176(9615/9984)
Starting timing for: iteration 1
The time was 00:00:03.0220205 for: iteration 1
Training: 0.0641092822944714
Testing: 0.0984236721922524(9667/9984)
Starting timing for: iteration 2
The time was 00:00:03.0370393 for: iteration 2
Training: 0.0310500149511131
Testing: 0.0893891200019145(9709/9984)
Starting timing for: iteration 3
The time was 00:00:03.0255130 for: iteration 3
Training: 0.0141359060303991
Testing: 0.0861167916184805(9734/9984)
Starting timing for: iteration 4
The time was 00:00:03.0400703 for: iteration 4
Training: 0.00617090886151299
Testing: 0.0828975517267678(9751/9984)
```

Now that I put it in it is slower. I guess at this size it is not enough to saturate the GPU.

```
                inl reproject a b ret =
                    inb x = s.CudaBlas.symm .Left .Lower one a.basic b.basic |> CudaAux.temporary
                    ret x

                inl reproject_to a b c = s.CudaBlas.symm' .Right .Lower -learning_rate b.basic a.basic one c.basic
```

Let me back it up here so I can undo and run it again."