"``` Starting timing for: iteration 0 The time was 00:00:02.9072668 for: iteration 0 Training: 0.1836399253414 Testing: 0.113177325552664(9620/9984) Starting timing for: iteration 1 The time was 00:00:02.8992085 for: iteration 1 Training: 0.0643514293120601 Testing: 0.095317505673171(9671/9984) Starting timing for: iteration 2 The time was 00:00:02.9057135 for: iteration 2 Training: 0.0305183720871961 Testing: 0.0884926123858(9712/9984) Starting timing for: iteration 3 The time was 00:00:02.9058609 for: iteration 3 Training: 0.0138830112601316 Testing: 0.0851460069254524(9749/9984) Starting timing for: iteration 4 The time was 00:00:02.9001392 for: iteration 4 Training: 0.00584066336190688 Testing: 0.0820976331742713(9754/9984) ```
This is using gemm. Let me switch to symm again.

```
Starting timing for: iteration 0
The time was 00:00:03.0410131 for: iteration 0
Training: 0.183482415679619
Testing: 0.115261024058176(9615/9984)
Starting timing for: iteration 1
The time was 00:00:03.0256886 for: iteration 1
Training: 0.0641092822944714
Testing: 0.0984236721922524(9667/9984)
Starting timing for: iteration 2
The time was 00:00:03.0212851 for: iteration 2
Training: 0.0310500149511131
Testing: 0.0893891200019145(9709/9984)
Starting timing for: iteration 3
The time was 00:00:03.0230783 for: iteration 3
Training: 0.0141359060303991
Testing: 0.0861167916184805(9734/9984)
Starting timing for: iteration 4
The time was 00:00:03.0211701 for: iteration 4
Training: 0.00617090886151299
Testing: 0.0828975517267678(9751/9984)
```

So it was not a fluke. Using `symm` does lead to a 4% slowdown for some reason.

5:05pm.

```
                inl reproject precision b ret =
                    inb x = s.CudaBlas.symm .Left .Upper one precision.basic b.basic |> CudaAux.temporary
                    ret x

                inl reproject_to a precision c = s.CudaBlas.symm' .Right .Upper -learning_rate precision.basic a.basic one c.basic
```

I give up. I'll undo. I guess symmetric multiplies lead to warp divergence.

It is good that I did this test. Rather than using a symmetric multiply for the raw factor, what I need is the triangular multiply `trmm` instead. I'll use that for weight sampling. It will be wonderful.

5:10pm. Oh, yes, I said I would bring back the regular init. Let me do that.

```
    inl mapi f in s =
        inl dim = Tensor.assert_broadcastable in
        inl in = to_dev_tensor in
        open CudaAD
        segmented_init {dim} (inl cur -> link_auto dim in cur >>= f cur) s

    inl map = mapi << const
```

No actually, I do not think there will be any problems with this. There is no need to bring back the regular `init`.

5:15pm. With this, I think I am done for the day. I certainly do not feel like starting anything else.

Implementing weight sampling would require at least two hours and my full focus during that duration. But apart from that it will be quite trivial to implement and test. I think in two days I'll at least be able to report from KFAC + weight sampling works on Mnist. Hopefully it will be something like 99% during the first iteration. I'll decrease the minibatch size to 32 to get it to sample more.

If something like that happens, then I'll know with absolute certainty where I stand. Nothing will be able to challenge me as I go to RL - nothing!

It is really a great dream.

And it is not just that. If I can get this to go my way, I will get a great dose of truth about what ML is trully all about. It will be an insight that nobody yet possesses."