I've been gathering my determination throughout the night and now is the time to let it lose. I can slack later. Let me just test this out. Once I try it on Mnist, I will know well enough where I stand with regard to my future.
Let me move to a separate branch called 'weight_sampling'. If this works, I won't be using standard nets anymore so I'll make it a firm change. I do not want Spiral's ML library to be for various different architectures - I want it to be for the sake of mastering a single good one.