"10:30am.
```
inl weight_sample {mult weight front={sampling} back={sampling}} =
```

...

```
    inl weight_sample {stddev weight front back} =
        inl dim = weight.basic.dim
        inl random = s.CudaRandom.create {stddev dst=.Normal; mean=0f32} {elem_type=float; dim}
        weight + (front.sampling * random * back.sampling)
```

Let me do it like this. I'll do the memory optimizations later. Let me implement the two operators.

```
    inl weight_sample {stddev weight front back} =
        inl dim = weight.basic.dim
        inl random = s.CudaRandom.create {stddev dst=.Normal; mean=0f32} {elem_type=float; dim}
        inl (*) a b = s.CudaBlas.gemm .nT .nT one a.basic b.basic |> View.wrap dim
        inl (+) a b = s.CudaBlas.geam .nT .nT one a.basic b.basic |> View.wrap dim

        weight + (front.sampling * random * back.sampling)
```

Screw triangular multiplies for now. That having said, am I sure that back does not have to be transposed or something? This is not an issue with inverse covariance as it is symmetric, but it might be an issue here. How complicated.

10:45am.

```
    inl weight_sample {d with stddev weight front back} =
        inl dim = weight.basic.dim
        inl weight = View.unzip weight
        inl random = s.CudaRandom.create {stddev dst=.Normal; mean=0f32} {elem_type=float; dim}
        inl (*) a b = s.CudaBlas.gemm .nT .nT one a.basic b.basic |> View.wrap dim
        inl (+) a b = s.CudaBlas.geam .nT .nT one a.basic b.basic |> View.wrap dim

        {
        d with weight = {
            primal = weight.primal + (front.sampling * random * back.sampling)
            adjoint = weight.adjoint
            block=()
            }
        }
```

How about this?

There is definitely a case to be made that back should be transposed though.

Well, let me get it to compile first and then I'll worry about the rest. Forget about getting top performance for the time being. I'll use low weight sampling noise and just get it to run first. Once I determine what should be transposed and what should not, then I will make a move to optimize all of this.

Let me modify the initializer."