"``` Starting timing for: iteration 0 The time was 00:00:03.0320894 for: iteration 0 Training: 0.800398222608571 Testing: 0.415826172269403(9267/9984) Starting timing for: iteration 1 The time was 00:00:02.9872826 for: iteration 1 Training: 0.347888070150203 Testing: 0.382842267404228(9400/9984) Starting timing for: iteration 2 The time was 00:00:02.9967793 for: iteration 2 Training: 0.235642283794452 Testing: 0.301400866274469(9524/9984) Starting timing for: iteration 3 The time was 00:00:02.9909756 for: iteration 3 Training: 0.178168402358617 Testing: 0.309436112135397(9541/9984) Starting timing for: iteration 4 The time was 00:00:02.9833607 for: iteration 4 Training: 0.138617502187645 Testing: 0.278205733501791(9564/9984) ```
Let me increase the noise level to `1 / sqrt a`.

```
Starting timing for: iteration 0
The time was 00:00:00.9451635 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

That rapidly gives me a nan. I am using the ln_relu activation so it should not be saturation that is the cause of this. Let me try transposing the back.

11:25am.

```
Starting timing for: iteration 0
The time was 00:00:03.0550799 for: iteration 0
Training: 0.804891127002481
Testing: 0.446462839559759(9245/9984)
Starting timing for: iteration 1
The time was 00:00:02.9939286 for: iteration 1
Training: 0.351213503059828
Testing: 0.346813293596661(9428/9984)
Starting timing for: iteration 2
The time was 00:00:03.0099547 for: iteration 2
Training: 0.238674096023474
Testing: 0.313274178067495(9504/9984)
Starting timing for: iteration 3
The time was 00:00:03.0031934 for: iteration 3
Training: 0.187386477488522
Testing: 0.291741896857648(9534/9984)
Starting timing for: iteration 4
The time was 00:00:03.0054104 for: iteration 4
Training: 0.146035718841186
Testing: 0.282257056572369(9558/9984)
```

It does not seem like transposing the back makes a lot of difference. I can't seem to take advantage of the full noise level without it blowing up.

Hmmmm...

It bothers me that the initializer gave errors. What would happen if I set stddev to zero?

Let me also initialize the means to something.

```
Starting timing for: iteration 0
The time was 00:00:02.9661087 for: iteration 0
Training: 0.183783397916306
Testing: 0.108347592099259(9647/9984)
Starting timing for: iteration 1
The time was 00:00:02.9334723 for: iteration 1
Training: 0.0628759594326719
Testing: 0.0956119306263132(9696/9984)
Starting timing for: iteration 2
The time was 00:00:02.9349695 for: iteration 2
Training: 0.0296673233288492
Testing: 0.0886363810404407(9716/9984)
Starting timing for: iteration 3
The time was 00:00:02.9334135 for: iteration 3
Training: 0.0128108311195646
Testing: 0.0819570623339715(9745/9984)
Starting timing for: iteration 4
The time was 00:00:02.9333885 for: iteration 4
Training: 0.00543976672148953
Testing: 0.0777739135381312(9767/9984)
```

Ok, this is with relu and stddev of 0. Let me up the noise level gradually.

```
stddev = val (to float 0.01) //val (one / (to float (View.span a)))
```

Let me do it like so."