"11:35am.
```
Starting timing for: iteration 0
The time was 00:00:02.9769529 for: iteration 0
Training: 0.486358880825564
Testing: 0.247177122154822(9338/9984)
Starting timing for: iteration 1
The time was 00:00:02.9159413 for: iteration 1
Training: 0.208139255662045
Testing: 0.220088004383844(9476/9984)
Starting timing for: iteration 2
The time was 00:00:02.9134230 for: iteration 2
Training: 0.144647770474775
Testing: 0.173470721222633(9566/9984)
Starting timing for: iteration 3
The time was 00:00:02.9105081 for: iteration 3
Training: 0.107298276839881
Testing: 0.183725426304124(9592/9984)
Starting timing for: iteration 4
The time was 00:00:02.9113190 for: iteration 4
Training: 0.0872420960761849
Testing: 0.17642086625841(9624/9984)
```

Ok, I'll slowly increase the noise along with the learning rate and see where this gets me.

11:40am.

```
Starting timing for: iteration 0
The time was 00:00:02.9805421 for: iteration 0
Training: 0.473694144688451
Testing: 0.326846113261313(9220/9984)
Starting timing for: iteration 1
The time was 00:00:02.9243851 for: iteration 1
Training: 0.241173502772203
Testing: 0.311783207723728(9366/9984)
Starting timing for: iteration 2
The time was 00:00:02.9186339 for: iteration 2
Training: 0.169467441505816
Testing: 0.29762833543683(9459/9984)
Starting timing for: iteration 3
The time was 00:00:02.9248985 for: iteration 3
Training: 0.118671409752788
Testing: 0.303962634537572(9508/9984)
Starting timing for: iteration 4
The time was 00:00:02.9183652 for: iteration 4
Training: 0.0885748362210833
Testing: 0.305988834205961(9529/9984)
```

I've pushed up the epsilon to -6.

```
out = { d with weight = View.zip {(View.unzip weight) with primal = self + ({T=front.sampling} * random * back.sampling)} }
```

I seem to be getting better results now that I've done this. Transposing the front? Well, I guess that is an option.

Yeah, it is not working well. If only I could get it to work better...

Let me do it for a bit more and then I will take a break.

11:55am. No, I can't get this to work.

Yeah, it would have been so great if it spiked to 99% immediately, but it won't be that easy. In order to get a sense for it, it might be good if I implemented Badam first. Right now I am a bit unsure whether the projections I am doing are sound. Now that I think about it, how did it work when I was doing PRONG directly on the backward pass? Ah, there I was using inverse covariances.

Let me check out the PRONG paper...

file:///E:/Downloads/Metaheuristics/Higher%20Order%20Optimization/prong/A%20Neural%20Network%20model%20with%20Bidirectional%20Whitening.pdf

9/16. Ah, here the back needs a transpose. Meaning, in my own code, it is the front which should be transposed instead to account for the inputs coming from a different direction.

12:05pm. Let me stop here. I still have a while to go before I throw in the towel on this."