Right now I am deep in thought reading papers. I am not sure how to interpret the results that I got and am unsure in which direction to go in.
Edit: Should have pushed this an hour ago.

I am so deep in thought that I am not sure what I am doing.

I think it might be possible to get this to work, but I need to do a deep dive on this. Let me read the Adam paper.