"3:55pm.
```
Adam employs a preconditioner that adapts to the geometry of the data, since vbt is an approximation to the diagonal of the Fisher information matrix (Pascanu & Bengio, 2013); however, Adam’s preconditioner (like AdaGrad’s) is more conservative in its adaption than vanilla NGD by preconditioning with the square root of the inverse of the diagonal Fisher information matrix approximation.
```

Why the square root? If only I could understand this.

4:20pm. They say it makes is go better. There is not much in the way of reasoning I could find as to why this is. Nor why the v^(1/2) is passed into the distribution. The paper is quite loose on this point.

Hmmm...ok, I am slowly getting over my shock. I was hoping KFAC + weight sampling would be like rolling thunder, but whatever. I'll go with what I have. Maybe it would be interesting to combine KFAC with momentum, but it is not a priority.

Ultimately, I need to pick an exploration method and I should go with this. Either way I will get it going.

Let me do more tests."