"4:25pm. There is a change I am considering making - how about I stop passing the learning rate directly and instead do it through the context `s`?
Yeah, let me do it.

4:30pm.

```
    met update_covariance x cov s =
        inl learning_rate = s.data.learning_rate.covariance
        inl k = x.span_outer
        inl alpha = Math.pow (one - learning_rate) k
        inl beta = one - alpha
        s.CudaBlas.syrk' .Lower .T (beta / to float k) x.basic alpha cov.basic // symmetric rank-k update. (beta / to float k) * x * x^T + alpha * cov
```

Let me do it like this.

```
    inl cost =
        inl s = s.data_add {learning_rate={weight=learning_rate; covariance=learning_rate ** 0.85f32}}
        Timer.time_it (string_format "iteration {0}" i)
        <| inl _ ->
            train {
                data={input=train_images; label=train_labels}
                network
                final
                } s
```

And like this.

Now in weight sampling I'll just test for existence of this field and if it now, just return the weight means.

```
        | _ ->
            {
            out = d
            bck = const ()
            }
```

Beautiful. Let me give it a try.

I should experiment with different rates for covariance. Maybe I should have them update more slowly than the main net?

Surprisingly, it compiled on the first try."