"4:35pm. The context might be the way to unify the train and test functions. I really should have gotten to that earlier, maybe I'll do it once I get this to work to my satisfaction.
```
Training: 2.73045560997775
Testing: 2.30258512496948(1008/9984)
Starting timing for: iteration 1
The time was 00:00:02.9147837 for: iteration 1
Training: 2.74360502734144
Testing: 2.30258512496948(1008/9984)
```

So much for getting it right on the first try.

How is this possible?

4:45pm.

```
Starting timing for: iteration 0
The time was 00:00:03.0233057 for: iteration 0
Training: 0.349826283510138
Testing: 0.149981985762469(9575/9984)
Starting timing for: iteration 1
The time was 00:00:02.9426403 for: iteration 1
Training: 0.142878101739891
Testing: 0.11699737577482(9659/9984)
Starting timing for: iteration 2
The time was 00:00:02.9446532 for: iteration 2
Training: 0.0998752225264031
Testing: 0.112905739706875(9680/9984)
Starting timing for: iteration 3
The time was 00:00:02.9428870 for: iteration 3
Training: 0.0728929809716398
Testing: 0.105809236998959(9720/9984)
Starting timing for: iteration 4
The time was 00:00:02.9482556 for: iteration 4
Training: 0.0561332489027041
Testing: 0.107669360772259(9723/9984)
```

Ok, I fixed the error. It does indeed work better during testing without weight sampling.

Let me do a sanity check by putting it back in."