"Let me now mess with the LR for the covariance matrix. Also I want to mess with the noise level. I think that instead of setting it layerwise, what I should do instead is set the layerwise noise to 1 / size and then increase it globally through the context.
Let me give it a try.

```
stddev = val (one / (to float (View.span a)))
```

Ah, no wait, I meant `span b`. Whops.

```
Starting timing for: iteration 0
The time was 00:00:02.9752524 for: iteration 0
Training: 0.229861534366277
Testing: 0.129056622136163(9609/9984)
Starting timing for: iteration 1
The time was 00:00:02.9119741 for: iteration 1
Training: 0.0865297637771393
Testing: 0.105234646302564(9660/9984)
Starting timing for: iteration 2
The time was 00:00:02.9086496 for: iteration 2
Training: 0.0463045981417124
Testing: 0.104019857436353(9683/9984)
Starting timing for: iteration 3
The time was 00:00:02.9080232 for: iteration 3
Training: 0.0250088527727012
Testing: 0.098943943235962(9718/9984)
Starting timing for: iteration 4
The time was 00:00:02.9144349 for: iteration 4
Training: 0.0135578588105534
Testing: 0.102374135632533(9727/9984)
```

This is not too bad now. Let me keep going."