"4:55pm.
```
stddev = val (one / sqrt (to float (View.span b)))
```

Oh, this level of noise works now.

```
Starting timing for: iteration 0
The time was 00:00:02.9892461 for: iteration 0
Training: 0.572508191341598
Testing: 0.199523438843057(9553/9984)
Starting timing for: iteration 1
The time was 00:00:02.9111618 for: iteration 1
Training: 0.257358540239063
Testing: 0.14731466822018(9673/9984)
Starting timing for: iteration 2
The time was 00:00:02.9050389 for: iteration 2
Training: 0.180346126338312
Testing: 0.150951370698632(9697/9984)
Starting timing for: iteration 3
The time was 00:00:02.9091157 for: iteration 3
Training: 0.140118508810515
Testing: 0.1445926230843(9700/9984)
Starting timing for: iteration 4
The time was 00:00:02.9092986 for: iteration 4
Training: 0.111194087204212
Testing: 0.136736499351103(9737/9984)
```

I need to mess with the LR for the covariance matrix.

...Did some renaming of `learning_rate` to `rate`.

```
Starting timing for: iteration 0
The time was 00:00:02.9664790 for: iteration 0
Training: 0.551691536303076
Testing: 0.228419860925687(9486/9984)
Starting timing for: iteration 1
The time was 00:00:02.9198695 for: iteration 1
Training: 0.241428827679229
Testing: 0.167764444327508(9617/9984)
Starting timing for: iteration 2
The time was 00:00:02.9425210 for: iteration 2
Training: 0.166012351032253
Testing: 0.153874009304486(9688/9984)
Starting timing for: iteration 3
The time was 00:00:02.9474104 for: iteration 3
Training: 0.125433647635609
Testing: 0.156861857766222(9684/9984)
Starting timing for: iteration 4
The time was 00:00:02.9309434 for: iteration 4
Training: 0.0942674036950459
Testing: 0.151904226363204(9701/9984)
```

Increasing the rate for the covariance matrix is not doing much. Right now the global rate is -8 which is quite high as it is. Let me increase the noise rate more.

```
Starting timing for: iteration 0
The time was 00:00:02.9619617 for: iteration 0
Training: 1.33509273975132
Testing: 0.376720297971268(9491/9984)
Starting timing for: iteration 1
The time was 00:00:02.9043017 for: iteration 1
Training: 0.589845209913806
Testing: 0.258124621965369(9648/9984)
Starting timing for: iteration 2
The time was 00:00:00.6201857 for: iteration 2
Training: NaN
Training diverged. Aborting...
```

Hmmm, it does not seem to be capable withstanding much more than I am already pressing it.

Let me try it with learning rates down by 2x."