"``` Starting timing for: iteration 0 The time was 00:00:02.9900259 for: iteration 0 Training: 1.62038765949571 Testing: 0.39344868587208(9416/9984) Starting timing for: iteration 1 The time was 00:00:02.9303913 for: iteration 1 Training: 0.703971176615192 Testing: 0.276238588781407(9598/9984) Starting timing for: iteration 2 The time was 00:00:02.9290347 for: iteration 2 Training: 0.502256718695036 Testing: 0.226432749959171(9684/9984) Starting timing for: iteration 3 The time was 00:00:02.9304843 for: iteration 3 Training: 0.407347736784663 Testing: 0.206970308764236(9696/9984) Starting timing for: iteration 4 The time was 00:00:02.6364768 for: iteration 4 Training: NaN Training diverged. Aborting... ```
Hmm, it does better with this now. Let me lower the LR once more.

```
Starting timing for: iteration 0
The time was 00:00:02.9544207 for: iteration 0
Training: 2.03084373668346
Testing: 0.536078747715193(9286/9984)
Starting timing for: iteration 1
The time was 00:00:02.9080613 for: iteration 1
Training: 0.889127077001473
Testing: 0.355052222015799(9484/9984)
Starting timing for: iteration 2
The time was 00:00:02.9079890 for: iteration 2
Training: 0.65739153079593
Testing: 0.280801838046424(9578/9984)
Starting timing for: iteration 3
The time was 00:00:02.9108027 for: iteration 3
Training: 0.54842128075707
Testing: 0.252435378648963(9615/9984)
Starting timing for: iteration 4
The time was 00:00:02.9090049 for: iteration 4
Training: 0.448839714758608
Testing: 0.230008165526033(9630/9984)
```

Let me run it for longer. Then I'll try it with tanh instead of ln_relu.

```
Starting timing for: iteration 0
The time was 00:00:02.9677226 for: iteration 0
Training: 2.03084373668346
Testing: 0.536078747715193(9286/9984)
Starting timing for: iteration 1
The time was 00:00:02.9091449 for: iteration 1
Training: 0.889127077001473
Testing: 0.355052222015799(9484/9984)
Starting timing for: iteration 2
The time was 00:00:02.9087944 for: iteration 2
Training: 0.65739153079593
Testing: 0.280801838046424(9578/9984)
Starting timing for: iteration 3
The time was 00:00:02.9091558 for: iteration 3
Training: 0.54842128075707
Testing: 0.252435378648963(9615/9984)
Starting timing for: iteration 4
The time was 00:00:02.9137563 for: iteration 4
Training: 0.448839714758608
Testing: 0.230008165526033(9630/9984)
Starting timing for: iteration 5
The time was 00:00:02.9143409 for: iteration 5
Training: 0.404471857175748
Testing: 0.213729103906373(9672/9984)
Starting timing for: iteration 6
The time was 00:00:02.9093801 for: iteration 6
Training: 0.358122213134685
Testing: 0.200236491836491(9681/9984)
Starting timing for: iteration 7
The time was 00:00:02.9116485 for: iteration 7
Training: 0.329249317424459
Testing: 0.189511088141517(9703/9984)
Starting timing for: iteration 8
The time was 00:00:02.9141481 for: iteration 8
Training: 0.309619450699896
Testing: 0.183172580872364(9706/9984)
Starting timing for: iteration 9
The time was 00:00:02.9116436 for: iteration 9
Training: 0.273589993306254
Testing: 0.173087553124377(9741/9984)
Starting timing for: iteration 10
The time was 00:00:02.9099219 for: iteration 10
Training: 0.264537673611649
Testing: 0.176667704079077(9727/9984)
Starting timing for: iteration 11
The time was 00:00:02.9104198 for: iteration 11
Training: 0.23710992918004
Testing: 0.172646328654157(9730/9984)
Starting timing for: iteration 12
The time was 00:00:02.9111136 for: iteration 12
Training: 0.223043692189716
Testing: 0.174999754263752(9730/9984)
Starting timing for: iteration 13
The time was 00:00:02.9126937 for: iteration 13
Training: 0.201261386701355
Testing: 0.170571095696954(9732/9984)
Starting timing for: iteration 14
The time was 00:00:02.9110296 for: iteration 14
Training: 0.196483932064625
Testing: 0.166114221006739(9742/9984)
```

I am actually interested in how far it can go so let me run it for even longer. It is interesting that now training is higher than test. I think that NNs should be in this regime generally."