"``` Starting timing for: iteration 15 The time was 00:00:01.2246113 for: iteration 15 Training: NaN Training diverged. Aborting... ```
Or maybe not. Let me try tanh.

```
Starting timing for: iteration 0
The time was 00:00:02.9699999 for: iteration 0
Training: 0.98516304896046
Testing: 0.351781501457224(9274/9984)
Starting timing for: iteration 1
The time was 00:00:02.9336320 for: iteration 1
Training: 0.562853319220181
Testing: 0.271094500069292(9429/9984)
Starting timing for: iteration 2
The time was 00:00:02.9254007 for: iteration 2
Training: 0.434043208988877
Testing: 0.223451497869041(9528/9984)
Starting timing for: iteration 3
The time was 00:00:02.9223301 for: iteration 3
Training: 0.354130065472574
Testing: 0.193626118277816(9600/9984)
Starting timing for: iteration 4
The time was 00:00:02.9239561 for: iteration 4
Training: 0.294751488825537
Testing: 0.189676388600732(9608/9984)
Starting timing for: iteration 5
The time was 00:00:02.9409906 for: iteration 5
Training: 0.257266041579951
Testing: 0.178404578430241(9614/9984)
Starting timing for: iteration 6
The time was 00:00:02.9221993 for: iteration 6
Training: 0.230071863739126
Testing: 0.173640355935497(9652/9984)
Starting timing for: iteration 7
The time was 00:00:02.9295555 for: iteration 7
Training: 0.198114386431472
Testing: 0.17565888839273(9642/9984)
Starting timing for: iteration 8
The time was 00:00:02.9256480 for: iteration 8
Training: 0.18210582624903
Testing: 0.168104228324861(9671/9984)
Starting timing for: iteration 9
The time was 00:00:02.9225986 for: iteration 9
Training: 0.157588195317658
Testing: 0.176505253965012(9676/9984)
Starting timing for: iteration 10
The time was 00:00:02.9267430 for: iteration 10
Training: 0.141365920044243
Testing: 0.16993326614816(9674/9984)
Starting timing for: iteration 11
The time was 00:00:02.9248297 for: iteration 11
Training: 0.129427055533752
Testing: 0.167298814690214(9684/9984)
Starting timing for: iteration 12
The time was 00:00:02.9305743 for: iteration 12
Training: 0.1140045483497
Testing: 0.170579054627958(9694/9984)
Starting timing for: iteration 13
The time was 00:00:02.9236076 for: iteration 13
Training: 0.107773609164879
Testing: 0.175721589473263(9696/9984)
Starting timing for: iteration 14
The time was 00:00:02.9253368 for: iteration 14
Training: 0.0892819950860131
Testing: 0.180788075139651(9697/9984)
Starting timing for: iteration 15
The time was 00:00:02.9233927 for: iteration 15
Training: 0.0897541411114873
Testing: 0.180190758007066(9689/9984)
Starting timing for: iteration 16
The time was 00:00:02.9247250 for: iteration 16
Training: 0.0840564467661552
Testing: 0.177003294311849(9706/9984)
Starting timing for: iteration 17
The time was 00:00:02.9239776 for: iteration 17
Training: 0.0730435875555128
Testing: 0.177725632579145(9706/9984)
Starting timing for: iteration 18
The time was 00:00:02.9261568 for: iteration 18
Training: 0.0673106390321874
Testing: 0.184626415768155(9707/9984)
Starting timing for: iteration 19
The time was 00:00:02.9242486 for: iteration 19
Training: 0.0668055383004965
Testing: 0.185561555208504(9686/9984)
Starting timing for: iteration 20
The time was 00:00:02.9268494 for: iteration 20
Training: 0.0586501296291426
Testing: 0.189717206172622(9692/9984)
Starting timing for: iteration 21
The time was 00:00:02.9309351 for: iteration 21
Training: 0.0601124286069737
Testing: 0.190519935947332(9705/9984)
Starting timing for: iteration 22
The time was 00:00:02.9312690 for: iteration 22
Training: 0.0523721258786336
Testing: 0.198182872021433(9699/9984)
Starting timing for: iteration 23
The time was 00:00:02.9369565 for: iteration 23
Training: 0.0517748132324239
Testing: 0.195298237460249(9698/9984)
Starting timing for: iteration 24
The time was 00:00:02.9484524 for: iteration 24
Training: 0.0482961260566106
Testing: 0.199252185832858(9704/9984)
Starting timing for: iteration 25
The time was 00:00:02.9317514 for: iteration 25
Training: 0.0446497444462476
Testing: 0.202491217518473(9710/9984)
Starting timing for: iteration 26
The time was 00:00:02.9457296 for: iteration 26
Training: 0.0398565841987641
Testing: 0.201495429676793(9704/9984)
Starting timing for: iteration 27
The time was 00:00:02.9269650 for: iteration 27
Training: 0.0403415683147179
Testing: 0.198489626896809(9710/9984)
Starting timing for: iteration 28
The time was 00:00:02.9303652 for: iteration 28
Training: 0.0358947838319695
Testing: 0.205502444354183(9692/9984)
Starting timing for: iteration 29
The time was 00:00:02.9269994 for: iteration 29
Training: 0.0366860443708762
Testing: 0.204772159012977(9713/9984)
```

This is with noise 0 and lr -10. The one that diverged previous was noise 1. Let me lower the noise to -1 and increase the lr to -9"