"```     inl network =         ln_relu 512,         ln_relu 512,         linear label_size ```
Let me try a larger network. This thing should be interesting.

```
Starting timing for: iteration 0
The time was 00:00:07.4110676 for: iteration 0
Training: 1.35811457294039
Testing: 0.502799867081098(9525/9984)
Starting timing for: iteration 1
The time was 00:00:07.3443443 for: iteration 1
Training: 0.497953487432793
Testing: 0.342500937320215(9689/9984)
Starting timing for: iteration 2
The time was 00:00:07.3393139 for: iteration 2
Training: 0.333111024837798
Testing: 0.285359304803102(9709/9984)
Starting timing for: iteration 3
The time was 00:00:07.3475776 for: iteration 3
Training: 0.259107676586322
Testing: 0.277164572845099(9705/9984)
Starting timing for: iteration 4
The time was 00:00:07.3539434 for: iteration 4
Training: 0.186618091005369
Testing: 0.24807880003918(9749/9984)
Starting timing for: iteration 5
The time was 00:00:07.3518678 for: iteration 5
Training: 0.156547633922148
Testing: 0.242699253071339(9753/9984)
Starting timing for: iteration 6
The time was 00:00:07.3584115 for: iteration 6
Training: 0.114291988272208
Testing: 0.232530852414082(9771/9984)
Starting timing for: iteration 7
The time was 00:00:07.3699506 for: iteration 7
Training: 0.0987721260653338
Testing: 0.244748437287733(9758/9984)
Starting timing for: iteration 8
The time was 00:00:07.3644914 for: iteration 8
Training: 0.0764191732081292
Testing: 0.239822899845868(9775/9984)
Starting timing for: iteration 9
The time was 00:00:07.3584625 for: iteration 9
Training: 0.0678877101256797
Testing: NaN(5358/5504)
Starting timing for: iteration 10
The time was 00:00:00.9619582 for: iteration 10
Training: NaN
Training diverged. Aborting...
```

Let me lower the noise.

5:35pm.

```
Starting timing for: iteration 0
The time was 00:00:07.4039715 for: iteration 0
Training: 0.60012363214404
Testing: 0.288261236291609(9557/9984)
Starting timing for: iteration 1
The time was 00:00:07.3723515 for: iteration 1
Training: 0.24080953667011
Testing: 0.237373932899707(9655/9984)
Starting timing for: iteration 2
The time was 00:00:07.3691751 for: iteration 2
Training: 0.146694712513255
Testing: 0.19971245941361(9710/9984)
Starting timing for: iteration 3
The time was 00:00:07.3875281 for: iteration 3
Training: 0.0868863748141265
Testing: 0.207228394716914(9715/9984)
Starting timing for: iteration 4
The time was 00:00:07.3472882 for: iteration 4
Training: 0.0584214456373677
Testing: 0.18414583323173(9752/9984)
Starting timing for: iteration 5
The time was 00:00:07.3656662 for: iteration 5
Training: 0.0398943648485482
Testing: 0.180930259502957(9752/9984)
Starting timing for: iteration 6
The time was 00:00:07.3672322 for: iteration 6
Training: 0.0271651081599095
Testing: 0.175844500882462(9759/9984)
Starting timing for: iteration 7
The time was 00:00:07.3541534 for: iteration 7
Training: 0.0203925494624877
Testing: 0.157427865117259(9775/9984)
Starting timing for: iteration 8
The time was 00:00:07.3538864 for: iteration 8
Training: 0.0161257205021696
Testing: 0.171637341345038(9784/9984)
Starting timing for: iteration 9
The time was 00:00:07.3653306 for: iteration 9
Training: 0.013771796963663
Testing: 0.163363348657877(9780/9984)
Starting timing for: iteration 10
The time was 00:00:07.3848251 for: iteration 10
Training: 0.0102532640981692
Testing: 0.147490379613604(9804/9984)
Starting timing for: iteration 11
The time was 00:00:07.4332829 for: iteration 11
Training: 0.00877297590232974
Testing: 0.15817106616118(9792/9984)
Starting timing for: iteration 12
The time was 00:00:07.3722870 for: iteration 12
Training: 0.00810663642536524
Testing: 0.160256002808029(9780/9984)
Starting timing for: iteration 13
The time was 00:00:07.3574119 for: iteration 13
Training: 0.00533800195289023
Testing: 0.169419503333735(9787/9984)
Starting timing for: iteration 14
The time was 00:00:07.3580386 for: iteration 14
Training: 0.00576617029260335
Testing: 0.15673659583149(9799/9984)
```

Let me jack up the noise and lower the lr. And I should consider removing sampling in the last layer."