"5:35pm.
```
Starting timing for: iteration 0
The time was 00:00:07.4039715 for: iteration 0
Training: 0.60012363214404
Testing: 0.288261236291609(9557/9984)
Starting timing for: iteration 1
The time was 00:00:07.3723515 for: iteration 1
Training: 0.24080953667011
Testing: 0.237373932899707(9655/9984)
Starting timing for: iteration 2
The time was 00:00:07.3691751 for: iteration 2
Training: 0.146694712513255
Testing: 0.19971245941361(9710/9984)
Starting timing for: iteration 3
The time was 00:00:07.3875281 for: iteration 3
Training: 0.0868863748141265
Testing: 0.207228394716914(9715/9984)
Starting timing for: iteration 4
The time was 00:00:07.3472882 for: iteration 4
Training: 0.0584214456373677
Testing: 0.18414583323173(9752/9984)
Starting timing for: iteration 5
The time was 00:00:07.3656662 for: iteration 5
Training: 0.0398943648485482
Testing: 0.180930259502957(9752/9984)
Starting timing for: iteration 6
The time was 00:00:07.3672322 for: iteration 6
Training: 0.0271651081599095
Testing: 0.175844500882462(9759/9984)
Starting timing for: iteration 7
The time was 00:00:07.3541534 for: iteration 7
Training: 0.0203925494624877
Testing: 0.157427865117259(9775/9984)
Starting timing for: iteration 8
The time was 00:00:07.3538864 for: iteration 8
Training: 0.0161257205021696
Testing: 0.171637341345038(9784/9984)
Starting timing for: iteration 9
The time was 00:00:07.3653306 for: iteration 9
Training: 0.013771796963663
Testing: 0.163363348657877(9780/9984)
Starting timing for: iteration 10
The time was 00:00:07.3848251 for: iteration 10
Training: 0.0102532640981692
Testing: 0.147490379613604(9804/9984)
Starting timing for: iteration 11
The time was 00:00:07.4332829 for: iteration 11
Training: 0.00877297590232974
Testing: 0.15817106616118(9792/9984)
Starting timing for: iteration 12
The time was 00:00:07.3722870 for: iteration 12
Training: 0.00810663642536524
Testing: 0.160256002808029(9780/9984)
Starting timing for: iteration 13
The time was 00:00:07.3574119 for: iteration 13
Training: 0.00533800195289023
Testing: 0.169419503333735(9787/9984)
Starting timing for: iteration 14
The time was 00:00:07.3580386 for: iteration 14
Training: 0.00576617029260335
Testing: 0.15673659583149(9799/9984)
```

Let me jack up the noise and lower the lr. And I should consider removing sampling in the last layer.

5:45pm.

```
Starting timing for: iteration 0
The time was 00:00:07.4272898 for: iteration 0
Training: 0.46877404673619
Testing: 0.306352269156774(9428/9984)
Starting timing for: iteration 1
The time was 00:00:07.3928898 for: iteration 1
Training: 0.218151027158362
Testing: 0.231160654859143(9577/9984)
Starting timing for: iteration 2
The time was 00:00:07.3725910 for: iteration 2
Training: 0.104603742643145
Testing: 0.204359077500414(9667/9984)
Starting timing for: iteration 3
The time was 00:00:07.3764349 for: iteration 3
Training: 0.0462883347504236
Testing: 0.190036224873488(9703/9984)
Starting timing for: iteration 4
The time was 00:00:07.3871440 for: iteration 4
Training: 0.0252777142733148
Testing: 0.174239767917886(9745/9984)
Starting timing for: iteration 5
The time was 00:00:07.3803221 for: iteration 5
Training: 0.0129550736817555
Testing: 0.168068530103984(9756/9984)
Starting timing for: iteration 6
The time was 00:00:07.3740587 for: iteration 6
Training: 0.00707479528623592
Testing: 0.160553190520175(9768/9984)
Starting timing for: iteration 7
The time was 00:00:07.3852324 for: iteration 7
Training: 0.00291820036180784
Testing: 0.150867274369046(9788/9984)
Starting timing for: iteration 8
The time was 00:00:07.3872631 for: iteration 8
Training: 0.000783458812751396
Testing: 0.144087566182672(9803/9984)
Starting timing for: iteration 9
The time was 00:00:07.3914552 for: iteration 9
Training: 0.000251680779420959
Testing: 0.143573017276954(9785/9984)
Starting timing for: iteration 10
The time was 00:00:07.3829095 for: iteration 10
Training: 0.00011736096289265
Testing: 0.140756427494845(9797/9984)
Starting timing for: iteration 11
The time was 00:00:07.3923096 for: iteration 11
Training: 4.35264838460789E-05
Testing: 0.138699777110258(9797/9984)
Starting timing for: iteration 12
The time was 00:00:07.3911316 for: iteration 12
Training: 2.77861178153878E-05
Testing: 0.137860005949833(9801/9984)
Starting timing for: iteration 13
The time was 00:00:07.3718077 for: iteration 13
Training: 1.87520715508169E-05
Testing: 0.137721492241438(9801/9984)
Starting timing for: iteration 14
The time was 00:00:07.3697011 for: iteration 14
Training: 2.37581875010252E-05
Testing: 0.137084981017717(9800/9984)
```

Tanh nets overfit a lot. I had to set noise to -3 to stop them from blowing up.

5:50pm.

```
Starting timing for: iteration 0
The time was 00:00:07.4330474 for: iteration 0
Training: 0.523186318914637
Testing: 0.305616292375362(9568/9984)
Starting timing for: iteration 1
The time was 00:00:07.3323087 for: iteration 1
Training: 0.193773912839936
Testing: 0.239841713572731(9647/9984)
Starting timing for: iteration 2
The time was 00:00:07.3254822 for: iteration 2
Training: 0.0784469467260314
Testing: 0.208978438993276(9697/9984)
Starting timing for: iteration 3
The time was 00:00:07.3455555 for: iteration 3
Training: 0.0397165229915117
Testing: 0.204663403217474(9714/9984)
Starting timing for: iteration 4
The time was 00:00:07.3455595 for: iteration 4
Training: 0.0213219665741274
Testing: 0.171711973308137(9765/9984)
Starting timing for: iteration 5
The time was 00:00:07.3344142 for: iteration 5
Training: 0.0116269622897363
Testing: 0.159398061164149(9785/9984)
Starting timing for: iteration 6
The time was 00:00:07.3289676 for: iteration 6
Training: 0.00580885296947258
Testing: 0.140541458438642(9799/9984)
Starting timing for: iteration 7
The time was 00:00:07.3278100 for: iteration 7
Training: 0.00230095739067492
Testing: 0.156616333721865(9786/9984)
Starting timing for: iteration 8
The time was 00:00:07.3471488 for: iteration 8
Training: 0.000471746435499055
Testing: 0.138228890517832(9808/9984)
Starting timing for: iteration 9
The time was 00:00:07.3510650 for: iteration 9
Training: 0.000122740214832216
Testing: 0.138717759758735(9799/9984)
Starting timing for: iteration 10
The time was 00:00:07.3420016 for: iteration 10
Training: 2.37564564178889E-05
Testing: 0.138499770681472(9804/9984)
Starting timing for: iteration 11
The time was 00:00:07.3366106 for: iteration 11
Training: 2.22116714330126E-05
Testing: 0.138077182659457(9806/9984)
Starting timing for: iteration 12
The time was 00:00:07.3322608 for: iteration 12
Training: 1.64431706374988E-05
Testing: 0.137483125160713(9805/9984)
Starting timing for: iteration 13
The time was 00:00:07.3330441 for: iteration 13
Training: 1.33098594390558E-05
Testing: 0.137348527318243(9805/9984)
Starting timing for: iteration 14
The time was 00:00:07.3331201 for: iteration 14
Training: 1.25055694454265E-05
Testing: 0.137138361297906(9807/9984)
```

This is ln_relu with ...

```
inl learning_rate = epsilon -8
inl pars = {rate={weight=learning_rate; covariance=learning_rate ** 0.85f32; noise=epsilon -3}}
Console.writeline pars
```

Seriously, I will just print out the damn parameters from here. Let me set noise to 0 and remove that sqrt. It does not seem like high noise is a virtue. It actually leeds to the net diverging and overfitting in strange ways."